Fix usage of removed method AppState.removeEventListener
AppState.removeEventListener was removed in react-native 0.70, after being
deprecated for over a year. Anybody using react-native 0.70 will experience an
unhandled TypeError when trying to invoke removeEventListener.

Replace the usage of removeEventListener with the `remove()` method on the
subscription returned by addEventListener.

Stop reassigning the resumeListener argument. It's a bad practice to reassign
function arguments, and it is common to have an eslint rule to block this.
Rename the argument to retryOnAppResume so that when we create resumeListener,
there is no reassignment of a function argument.